question2/aws_batch__michaelkors_scrape/requirements.txt
aioboto3==9.4.0
aiobotocore==2.1.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
boto3==1.20.24
botocore==1.23.24
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
distlib==0.3.4
filelock==3.6.0
frozenlist==1.3.0
h11==0.13.0
html5lib==1.1
idna==3.3
iniconfig==1.1.1
jmespath==0.10.0
lxml==4.8.0
multidict==6.0.2
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.1
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.17.1
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
webdriver-manager==3.5.4
webencodings==0.5.1
wrapt==1.14.0
wsproto==1.1.0
yarl==1.7.2
